refactor(av1)!: widen reconstruction to u16 for high-bit-depth support
Make the reconstruction path bit-depth-generic so the M2 10/12-bit work can
slot in, with **no change to 8-bit output** (every dav1d/libavif bit-exact
test stays byte-identical):

- gamut-color: add `clip_pixel(x, bit_depth) -> u16` (AV1 `Clip1`), the
  high-bit-depth companion to `clip_pixel8`.
- gamut-av1: the `FrameEncoder` reconstruction buffer, `Reconstruction`, and
  the in-loop filters (deblock/CDEF/loop-restoration/superres-upscale) now hold
  `u16` samples; a `bit_depth` field threads through the clamp and the
  dc_q/ac_q lookups. The filter *math* stays 8-bit (the `±128` centring etc.);
  bit-depth scaling lands with the 10-bit path.
- The dev-only `dav1d-oracle`/`libavif-oracle` decode 8/10/12-bit pictures,
  returning `u16` planes plus a `bit_depth` (previously they hard-rejected
  anything but 8-bit) — the prerequisite for testing high-bit-depth output.

BREAKING CHANGE: `gamut_av1::ReconImage::planes` is now `[Vec<u16>; 3]` and the
struct gains a `bit_depth: u8` field.

just test (recon_dav1d 26, decode_roundtrip, + units) green; lint/format clean.

+++ crates/gamut-avif/tests/decode_roundtrip.rs
@@ -59,9 +59,9 @@ fn roundtrip(w: u32, h: u32) {
         for x in 0..w {
             let i = (y * w + x) as usize;
             let (r, g, b) = rgb_at(x, y);
-            assert_eq!(yp[i], g, "Y!=G at ({x},{y}) in {w}x{h}");
-            assert_eq!(up[i], b, "U!=B at ({x},{y}) in {w}x{h}");
-            assert_eq!(vp[i], r, "V!=R at ({x},{y}) in {w}x{h}");
+            assert_eq!(yp[i], u16::from(g), "Y!=G at ({x},{y}) in {w}x{h}");
+            assert_eq!(up[i], u16::from(b), "U!=B at ({x},{y}) in {w}x{h}");
+            assert_eq!(vp[i], u16::from(r), "V!=R at ({x},{y}) in {w}x{h}");
         }
     }
 }
@@ -162,9 +162,13 @@ fn orientation_transforms_roundtrip_via_libavif() {
             for x in 0..w {
                 let i = (y * w + x) as usize;
                 let (r, g, b) = rgb_at(x, y);
-                assert_eq!(yp[i], g, "Y at ({x},{y}) irot={rot} imir={mir:?}");
-                assert_eq!(up[i], b);
-                assert_eq!(vp[i], r);
+                assert_eq!(
+                    yp[i],
+                    u16::from(g),
+                    "Y at ({x},{y}) irot={rot} imir={mir:?}"
+                );
+                assert_eq!(up[i], u16::from(b));
+                assert_eq!(vp[i], u16::from(r));
             }
         }
     }
